requirements.txt
face_recognition
opencv-python
pillow 
imgaug
faiss-cpu
mtcnn
gradio
tensorflow